refactor: update direct API services for new architecture
- Update calendar service to use DragonflyDB cache
- Update weather service with new decorators and cache
- Update flights service imports and structure
- All services now use direct SDK integration
