+++ tripsage/services/api/weather_service.py
@@ -24,7 +24,7 @@ from tripsage.models.api.weather_models import (
     WeatherUnits,
 )
 from tripsage.services.base import BaseService
-from tripsage.utils.cache_tools import cached
+from tripsage.utils.cache import cached
 from tripsage.utils.decorators import with_retry
 from tripsage.utils.logging import get_logger
